Refactor SQL to use INT for DepartmentID

+++ Assignment01_Create_Department/solution.sql
@@ -10,7 +10,7 @@ CREATE DATABASE CollegeDB;
 USE CollegeDB;
 
 CREATE TABLE Department (
-    DepartmentID NUMBER(5) PRIMARY KEY,
+    DepartmentID INT PRIMARY KEY,
     DepartmentName VARCHAR(20),
     HOD VARCHAR(20)
 );
